refactor: App 모듈 생성 & UI 책임 분리
